requirements.txt
scikit-learn==1.1.3
scikit-image==0.19.3
numpy>=1.23.5
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
matplotlib==3.6.2
streamlit